assignment 5 is now solved
Problem 4 required the use of hashcode and equals for the entry class in order for everything to technically work. Everything else was kinda straight forward because all you were doing was kinda reusing code from the QuadraticProbingHashTable Class. The get method was the only interesting one because you don't fill in a ValueType when you are searching for the entry you want with your key